Relocate Immutability Enforcement Tests
These tests are of runner behavior rather than the model, and should be
tested as a runner test.

Stop wrapping IllegalMutationExceptions to surface failures due to
model invariant violations directly, rather than going through a
PipelineExecutionException.
